+++ Changelog.md
@@ -3,9 +3,12 @@
 - New: Reinforced Blocks: Pale Oak Log, Stripped Pale Oak Log, Pale Oak Wood, Stripped Pale Oak Wood, Pale Oak Planks, Pale Oak Stairs, Pale Oak Slabs, Pale Oak Button, Pale Oake Pressure Plate, Pale Oak Fence, Pale Oak Fence Gate, Resin Block, Resin Bricks, Chiseled Resin Bricks, Resin Brick Slab, Resin Brick Stairs, Resin Brick Wall, Pale Moss Block, Pale Moss Carpet
 - New: Secret Pale Oak (Hanging) Sign
 - New: Pale Oak Security Sea Boat
+- New: Creaking Heart Mine
 - New: Cameras can now be viewed on Frames within the world (live camera feeds)[1]
 - New: Several server and client config settings to control chunk loading in camera feeds
 - Change: Camera model animations are now synchronized between players
+- API: New IBlockMine interface for blocks that are block mines
+- API: New method IExplosive#explodesWhenInteractedWith as well as two utility methods
 - Fix: The display of items in the SC Manual can change too fast in certain situations
 - Fix: Potential startup crash